# increase Transaction.tx_id performance