Done
- use iolist to minimize binary concatenation in Transaction serialization function(https://hexdocs.pm/elixir/IO.html#module-use-cases-for-io-data)
- TxUtils.serialize_legacy to return iolist instead of hex encoded string
- better typing
- add benchee as dev deps and one sample script of using benchee for benchmark

-----
Before using iolist,

```
Name                                    ips        average  deviation         median         99th %
&Transaction.transaction_id/1        4.81 K      207.94 μs    ±10.61%         202 μs         305 μs

Memory usage statistics:

Name                             Memory usage
&Transaction.transaction_id/1        14.16 KB
```

After using iolist,

```
Name                                    ips        average  deviation         median         99th %
&Transaction.transaction_id/1        5.12 K      195.20 μs     ±9.75%         190 μs         279 μs

Memory usage statistics:

Name                             Memory usage
&Transaction.transaction_id/1        14.26 KB
```

After removing unnecessary Base16.encode in `TxUtils.serialize_legacy`  

```
Name                                    ips        average  deviation         median         99th %
&Transaction.transaction_id/1       11.55 K       86.60 μs    ±16.44%          83 μs         152 μs

Memory usage statistics:

Name                             Memory usage
&Transaction.transaction_id/1        11.72 KB
```